[TT-021/001] Made the list view as main window, Changed colour of the window [EP, SS]
