+++ TodoTasker/app/src/main/res/layout/activity_main.xml
@@ -1,93 +1,23 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- No Hard-Coded Strings or values. Define everything in the "values" folder corresponding file-->
-<!-- Default Layout should be suited for a resolution of 1080x2220 -->
-<!-- IDs: -->
-<!--        Naming Convention: elementtype_usage_with_short_desc -->
-<!-- Constants(Strings): -->
-<!--        Naming Convention: TYPE_USAGESHORTDESC-->
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".MainActivity">
+    <!-- No Hard-Coded Strings or values. Define everything in the "values" folder corresponding file-->
+    <!-- Default Layout should be suited for a resolution of 1080x2220 -->
+    <!-- IDs: -->
+    <!--        Naming Convention: elementtype_usage_with_short_desc -->
+    <!-- Constants(Strings): -->
+    <!--        Naming Convention: TYPE_USAGESHORTDESC-->
 
-    <TextView
-        android:id="@+id/textview_add_task_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/STRING_CREATETASK"
-        android:textSize="44sp"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/todo_list"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.023" />
-
-    <Button
-        android:id="@+id/button_add_to_db"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="150dp"
-        android:layout_marginBottom="40dp"
-        android:backgroundTint="#4CAF50"
-        android:onClick="addTodoActivity"
-        android:text="@string/STRING_ADDTASK"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/edittext_name"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <EditText
-        android:id="@+id/edittext_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="100dp"
-        android:layout_marginTop="184dp"
-        android:layout_marginEnd="100dp"
-        android:autofillHints=""
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:labelFor="@+id/edittext_name"
-        android:text="@string/STRING_NAME"
-        android:textAlignment="center"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textview_add_task_title" />
-
-    <Button
-        android:id="@+id/button_switch_to_list"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="44dp"
-        android:text="@string/STRING_SHOWTASKLIST"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button_add_to_db" />
-
-    <EditText
-        android:id="@+id/edittext_datetime"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="100dp"
-        android:layout_marginEnd="100dp"
-        android:clickable="true"
-        android:cursorVisible="false"
-        android:ems="10"
-        android:focusable="false"
-        android:inputType="datetime"
-        android:text="@string/STRING_ENTER_DATE"
-        android:textAlignment="center"
-        android:onClick="clickOnDateTimeField"
-        app:layout_constraintBottom_toTopOf="@+id/button_add_to_db"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/edittext_name" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+        />
+</LinearLayout>
